news_patch_kit_plus_collector/requirements.txt
pandas
numpy
scikit-learn
PyYAML
pymysql
